requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2==2.9.9
requests==2.31.0
Pillow==9.4.0
django-cors-headers
dj-rest-auth
django-allauth
psycopg2-binary  
dj-database-url  
whitenoise  
waitress  